Extract input handling
