+++ cmd/cli/cli.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/codeclimate/hestia/internal/commands"
 	"github.com/codeclimate/hestia/internal/notifiers"
-	"github.com/codeclimate/hestia/internal/types"
+	"github.com/codeclimate/hestia/internal/utils"
 	"log"
 	"os"
 	"os/user"
@@ -12,7 +12,8 @@ import (
 )
 
 func main() {
-	input := extractInput(strings.Join(os.Args[1:], " "))
+	re := regexp.MustCompile(`(?P<command>\w+)\s?(?P<args>.*)?`)
+	input := utils.ExtractInput(strings.Join(os.Args[1:], " "), re)
 
 	log.Printf("command = %s\n", input.Command)
 	log.Printf("args = %s\n", input.Args)
@@ -23,23 +24,4 @@ func main() {
 
 	command := commands.Build(user.Username, input, notifier)
 	command.Run()
-}
-
-func extractInput(text string) types.Input {
-	re := regexp.MustCompile(`(?P<command>\w+)\s?(?P<args>.*)?`)
-	match := re.FindStringSubmatch(text)
-	captures := extractCaptures(re, match)
-
-	return types.Input{Command: captures["command"], Args: captures["args"]}
-}
-
-func extractCaptures(re *regexp.Regexp, match []string) map[string]string {
-	captures := make(map[string]string)
-	for i, name := range re.SubexpNames() {
-		if i > 0 && i <= len(match) {
-			captures[name] = match[i]
-		}
-	}
-
-	return captures
 }
